requirements_cortex.txt
slack-bolt>=1.18.0
snowflake-connector-python>=3.0.0

